refactor: refactor geojson.py into several files
# Conflicts:
#	udata_hydra/analysis/geojson.py
